Implemented dates hierarchies rendering.
How does it work ?

To have it working, you need to activate the flag "render.dates" in config, then to define a data path using "dates.path" property (also in config).
Finally, the template.date.file need to be defined using the template engine of your choice.

Once all of that has been done, you can start templating your date-based rendering.

In this hierarchy, a special set of model data are available :

 - date_year contains the year this page will use as period reference. It should be more used as a marker than to render anything, as it is not formatted.
 - date_month contains the month this page will use as period reference. It should be more used as a marker than to render anything, as it is not formatted.
 - date_day contains the day this page will use as period reference. It should be more used as a marker than to render anything, as it is not formatted.
 - interval will contain the joda-tome object representing the current interval (notice period date can be obtained by interval.getStart().toDate()
 - published_subintervals will contain a map linking file names to the date for the start of the associated subperiod (for months, it will contain "01/index.html"=>"date of jan. 1st")
And, finally
 - published_during will contain a sorted list of all elements published during the given period.

Some things are not yet working, but I push it anyway for commenting.

What is not working ? well, relative link computation, as a starter (I have not yet merged all the changes I made in other branches).